+++ src/main/java/com/example/demo/customer/service/CustomerService.java
@@ -5,27 +5,27 @@ import java.util.List;
 import com.example.demo.customer.dto.*;
 
 public interface CustomerService {
-	CustomerDto getCustomerById(Long id);
+	CustomerDto getCustomerById(Integer id);
 
     CustomerDto createCustomer(CustomerCreateDto dto);
 
-    CustomerDto updateCustomer(Long id, CustomerUpdateDto dto);
+    CustomerDto updateCustomer(Integer id, CustomerUpdateDto dto);
 
-    AddressDto updateAddress(Long customerId, AddressDto dto);
+    AddressDto updateAddress(Integer customerId, AddressDto dto);
 
-    List<RentalResponseDto> getCustomerRentals(Long customerId);
+    List<RentalResponseDto> getCustomerRentals(Integer customerId);
 
-    List<PaymentDto> getCustomerPayments(Long customerId);
+    List<PaymentDto> getCustomerPayments(Integer customerId);
 
     List<CustomerDto> getCustomersByName(String name);
 
     List<CustomerDto> getCustomersByLocation(String location);
 
-    CustomerDto patchCustomer(Long id, CustomerUpdateDto dto);
+    CustomerDto patchCustomer(Integer id, CustomerUpdateDto dto);
 
-    AddressDto patchAddress(Long customerId, AddressDto dto);
+    AddressDto patchAddress(Integer customerId, AddressDto dto);
 
-    PaymentDto getCustomerPaymentById(Long customerId, Long paymentId);
+    PaymentDto getCustomerPaymentById(Integer customerId, Integer paymentId);
 
-    RentalResponseDto getCustomerRentalById(Long customerId, Long rentalId);
+    RentalResponseDto getCustomerRentalById(Integer customerId, Integer rentalId);
 }
